Start to refactor
